requirements.txt
numpy==1.18.5
pandas==1.0.5
matplotlib==3.3.1
scikit_learn==0.23.2
tqdm==4.48.2
librosa==0.8.0
torch==1.7.0+cu101
torchvision==0.8.1+cu101
pytorch_lightning==1.0.5
timm==0.3.1
hydra_core==1.0.4
omegaconf==2.0.5